Condense turf statistics sprawl into turf-collect
