--- packages/turf-aggregate/package.json
@@ -1,45 +1,0 @@
-{
-  "name": "turf-aggregate",
-  "version": "1.0.3",
-  "description": "turf aggregate module",
-  "main": "index.js",
-  "scripts": {
-    "test": "tape test.js",
-    "doc": "dox -r < index.js | doxme --readme > README.md"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/Turfjs/turf-aggregate.git"
-  },
-  "keywords": [
-    "turf",
-    "aggregate",
-    "points",
-    "geojson",
-    "polygons",
-    "stats"
-  ],
-  "author": "morganherlocker",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/Turfjs/turf-aggregate/issues"
-  },
-  "homepage": "https://github.com/Turfjs/turf-aggregate",
-  "devDependencies": {
-    "benchmark": "^1.0.0",
-    "tape": "^3.5.0",
-    "dox": "^0.6.1",
-    "doxme": "^1.4.3",
-    "turf-helpers": "^1.0.0"
-  },
-  "dependencies": {
-    "turf-average": "^1.1.0",
-    "turf-count": "^1.0.1",
-    "turf-deviation": "^1.0.0",
-    "turf-max": "^1.0.0",
-    "turf-median": "^1.0.1",
-    "turf-min": "^1.0.0",
-    "turf-sum": "^1.0.0",
-    "turf-variance": "^1.0.0"
-  }
-}
