--- packages/turf-average/package.json
@@ -1,38 +1,0 @@
-{
-  "name": "turf-average",
-  "version": "1.1.2",
-  "description": "turf average module",
-  "main": "index.js",
-  "scripts": {
-    "test": "tape test.js",
-    "doc": "dox -r < index.js | doxme --readme > README.md"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/Turfjs/turf-average.git"
-  },
-  "keywords": [
-    "turf",
-    "average",
-    "avg",
-    "stats",
-    "geojson",
-    "geospatial"
-  ],
-  "author": "morganherlocker",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/Turfjs/turf-average/issues"
-  },
-  "homepage": "https://github.com/Turfjs/turf-average",
-  "devDependencies": {
-    "benchmark": "^1.0.0",
-    "tape": "^3.5.0",
-    "dox": "^0.6.1",
-    "doxme": "^1.4.3",
-    "turf-helpers": "^1.0.0"
-  },
-  "dependencies": {
-    "turf-inside": "^1.1.3"
-  }
-}
