+++ packages/turf-collect/test.js
@@ -1,12 +1,12 @@
-var max = require('./');
+var aggregate = require('./');
 var test = require('tape');
 var polygon = require('turf-helpers').polygon;
 var point = require('turf-helpers').point;
 var featurecollection = require('turf-helpers').featureCollection;
 
-test('max', function(t){
-  var poly1 = polygon([[[0,0],[10,0],[10,10], [0,10],[0,0]]]);
-  var poly2 = polygon([[[10,0],[20,10],[20,20], [20,0],[10,0]]]);
+test('aggregate', function(t){
+  var poly1 = polygon([[[0,0],[10,0],[10,10],[0,10],[0,0]]]);
+  var poly2 = polygon([[[10,0],[20,10],[20,20],[20,0],[10,0]]]);
   var polyFC = featurecollection([poly1, poly2]);
   var pt1 = point([5,5], {population: 200});
   var pt2 = point([1,3], {population: 600});
@@ -14,12 +14,8 @@ test('max', function(t){
   var pt4 = point([13,1], {population: 200});
   var pt5 = point([19,7], {population: 300});
   var ptFC = featurecollection([pt1, pt2, pt3, pt4, pt5]);
-
-  var maxed = max(polyFC, ptFC, 'population', 'pop_max');
-
-  t.equal(maxed.features[0].geometry.type, 'Polygon');
-  t.equal(maxed.features[0].properties.pop_max, 600);
-  t.equal(maxed.features[1].properties.pop_max, 300);
-
+  var aggregated = aggregate(polyFC, ptFC, 'population', 'values');
+  t.deepEqual(aggregated.features[0].properties.values, [200, 600]);
+  t.deepEqual(aggregated.features[1].properties.values, [100, 200, 300]);
   t.end();
-})
+});
