--- packages/turf-deviation/package.json
@@ -1,38 +1,0 @@
-{
-  "name": "turf-deviation",
-  "version": "1.0.1",
-  "description": "turf deviation module",
-  "main": "index.js",
-  "scripts": {
-    "test": "tape test.js",
-    "doc": "dox -r < index.js | doxme --readme > README.md"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/Turfjs/turf-deviation.git"
-  },
-  "keywords": [
-    "turf",
-    "geojson",
-    "standard",
-    "deviation",
-    "stats"
-  ],
-  "author": "morganherlocker",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/Turfjs/turf-deviation/issues"
-  },
-  "homepage": "https://github.com/Turfjs/turf-deviation",
-  "devDependencies": {
-    "benchmark": "^1.0.0",
-    "tape": "^3.5.0",
-    "dox": "^0.6.1",
-    "doxme": "^1.4.3",
-    "turf-helpers": "^1.0.0"
-  },
-  "dependencies": {
-    "simple-statistics": "^0.9.0",
-    "turf-inside": "^1.1.3"
-  }
-}
