--- packages/turf-min/package.json
@@ -1,36 +1,0 @@
-{
-  "name": "turf-min",
-  "version": "1.0.1",
-  "description": "turf min module",
-  "main": "index.js",
-  "scripts": {
-    "test": "tape test.js",
-    "doc": "dox -r < index.js | doxme --readme > README.md"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/Turfjs/turf-min.git"
-  },
-  "keywords": [
-    "turf",
-    "min",
-    "stats",
-    "geojson"
-  ],
-  "author": "morganherlocker",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/Turfjs/turf-min/issues"
-  },
-  "homepage": "https://github.com/Turfjs/turf-min",
-  "devDependencies": {
-    "benchmark": "^1.0.0",
-    "tape": "^3.5.0",
-    "dox": "^0.6.1",
-    "doxme": "^1.4.3",
-    "turf-helpers": "^1.0.0"
-  },
-  "dependencies": {
-    "turf-inside": "^1.1.3"
-  }
-}
