--- packages/turf-variance/package.json
@@ -1,38 +1,0 @@
-{
-  "name": "turf-variance",
-  "version": "1.0.1",
-  "description": "turf variance module",
-  "main": "index.js",
-  "scripts": {
-    "test": "tape test.js",
-    "doc": "dox -r < index.js | doxme --readme > README.md"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/Turfjs/turf-variance.git"
-  },
-  "keywords": [
-    "turf",
-    "geojson",
-    "variance",
-    "stats",
-    "features"
-  ],
-  "author": "morganherlocker",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/Turfjs/turf-variance/issues"
-  },
-  "homepage": "https://github.com/Turfjs/turf-variance",
-  "devDependencies": {
-    "benchmark": "^1.0.0",
-    "tape": "^3.5.0",
-    "dox": "^0.6.1",
-    "doxme": "^1.4.3",
-    "turf-helpers": "^1.0.0"
-  },
-  "dependencies": {
-    "simple-statistics": "^0.9.0",
-    "turf-inside": "^1.1.3"
-  }
-}
