+++ packages/turf/index.js
@@ -12,18 +12,10 @@ module.exports = {
     convex: require('turf-convex'),
     within: require('turf-within'),
     concave: require('turf-concave'),
-    count: require('turf-count'),
     erase: require('turf-erase'),
-    variance: require('turf-variance'),
-    deviation: require('turf-deviation'),
-    median: require('turf-median'),
-    min: require('turf-min'),
-    max: require('turf-max'),
-    aggregate: require('turf-aggregate'),
+    collect: require('turf-collect'),
     flip: require('turf-flip'),
     simplify: require('turf-simplify'),
-    sum: require('turf-sum'),
-    average: require('turf-average'),
     bezier: require('turf-bezier'),
     tag: require('turf-tag'),
     size: require('turf-size'),
