+++ packages/turf/package.json
@@ -49,7 +49,6 @@
     "turf-aggregate": "1.0.2",
     "turf-along": "^1.0.2",
     "turf-area": "^1.1.1",
-    "turf-average": "1.1.1",
     "turf-bbox-polygon": "^1.0.1",
     "turf-bearing": "1.0.1",
     "turf-bezier": "1.0.2",
@@ -59,9 +58,7 @@
     "turf-combine": "1.0.2",
     "turf-concave": "1.1.3",
     "turf-convex": "1.0.1",
-    "turf-count": "1.0.2",
     "turf-destination": "^1.2.1",
-    "turf-deviation": "1.0.1",
     "turf-distance": "^1.0.1",
     "turf-envelope": "~1.0.2",
     "turf-erase": "1.3.1",
@@ -76,10 +73,7 @@
     "turf-kinks": "^1.3.1",
     "turf-line-distance": "^1.0.2",
     "turf-line-slice": "^1.3.4",
-    "turf-max": "1.0.1",
-    "turf-median": "1.0.2",
     "turf-midpoint": "^1.0.1",
-    "turf-min": "1.0.1",
     "turf-nearest": "^1.0.2",
     "turf-planepoint": "1.0.1",
     "turf-point-grid": "^2.0.0",
@@ -93,12 +87,10 @@
     "turf-size": "^1.1.1",
     "turf-square": "^1.0.1",
     "turf-square-grid": "^1.0.0",
-    "turf-sum": "1.0.1",
     "turf-tag": "~1.0.1",
     "turf-tin": "1.0.6",
     "turf-triangle-grid": "^1.0.0",
     "turf-union": "1.0.1",
-    "turf-variance": "1.0.1",
     "turf-within": "^1.0.1",
     "turf-helpers": "^1.0.0"
   },
